requirements.txt
tensorflow>=2.17.0,<2.19.0
mediapipe>=0.10.8
opencv-python>=4.8.1.78
google-cloud-language>=2.13.3
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.3
pillow>=10.1.0
requests>=2.31.0
python-dotenv>=1.0.0

